requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.30
psycopg2-binary==2.9.9

celery==5.4.0
redis==5.0.4

python-multipart==0.0.9

qdrant-client==1.9.1

langchain==0.2.3
langchain-openai==0.1.8

pypdf==4.2.0
pytesseract==0.3.10
Pillow==10.3.0
python-docx==1.1.2

streamlit==1.36.0
requests==2.32.3
pdf2image
openai>=1.0.0

PyPDF2